refactor: Rename S3StageTable into StageTable
