+++ common/planners/src/lib.rs
@@ -47,8 +47,8 @@ mod plan_node_display;
 mod plan_node_display_indent;
 mod plan_node_extras;
 mod plan_node_rewriter;
-mod plan_node_s3_stage_table;
 mod plan_node_stage;
+mod plan_node_stage_table;
 mod plan_node_statistics;
 mod plan_node_visitor;
 mod plan_partition;
@@ -160,9 +160,9 @@ pub use plan_node_builder::PlanBuilder;
 pub use plan_node_extras::Extras;
 pub use plan_node_rewriter::PlanRewriter;
 pub use plan_node_rewriter::RewriteHelper;
-pub use plan_node_s3_stage_table::S3StageTableInfo;
 pub use plan_node_stage::StageKind;
 pub use plan_node_stage::StagePlan;
+pub use plan_node_stage_table::StageTableInfo;
 pub use plan_node_statistics::Statistics;
 pub use plan_node_visitor::PlanVisitor;
 pub use plan_partition::PartInfo;
