+++ query/src/storages/mod.rs
@@ -19,17 +19,15 @@ pub mod index;
 pub mod information_schema;
 pub mod memory;
 pub mod null;
+pub mod stage;
 pub mod system;
 pub mod view;
 
-mod s3;
 mod storage_context;
 mod storage_factory;
 mod storage_table;
 mod storage_table_read_plan;
 
-pub use s3::S3StageTable;
-pub use s3::StageSource;
 pub use storage_context::StorageContext;
 pub use storage_factory::StorageCreator;
 pub use storage_factory::StorageDescription;
